Fix S.13 addendum IPC QA follow-ups
