+++ crates/atm-daemon/src/tests.rs
@@ -1,17 +1,19 @@
 use super::runtime_health::{DaemonRequestDispatcher, RuntimeStatusCache};
 use super::{
-    LocalIpcServerTransportAdapter,
+    DaemonExitCode, LocalIpcServerTransportAdapter, daemon_exit_code_for_error,
     host_ownership::{
         HOST_RUNTIME_OWNER_LOCK_FILE, HostOwnershipAdapter, clear_stale_recovery_signal_for_test,
         install_stale_recovery_signal_for_test,
     },
     lifecycle_control::LifecycleControlSourceAdapter,
     local_ipc_transport::RuntimeServeHooks,
+    test_support::{LifecycleFlagResetGuard, connect_daemon_local_ipc_until_ready},
 };
 use atm_core::boundary::RequestDispatcher;
 #[cfg(unix)]
 use atm_core::doctor::DoctorQuery;
 use atm_core::doctor::{DoctorEnvironmentVisibility, DoctorReport, DoctorStatus, DoctorSummary};
+use atm_core::error::AtmError;
 use atm_core::error_codes::AtmErrorCode;
 use atm_core::observability::{AtmObservabilityHealth, AtmObservabilityHealthState};
 use atm_core::protocol::{
@@ -25,8 +27,6 @@ use atm_core::test_support::ROLE_TEAM_LEAD;
 use atm_core::types::{AgentName, IsoTimestamp, TeamName};
 use atm_rusqlite::assemble_boundary;
 #[cfg(unix)]
-use interprocess::local_socket::Stream as LocalSocketStream;
-#[cfg(unix)]
 use interprocess::local_socket::traits::Stream as _;
 use serial_test::serial;
 use std::fs::OpenOptions;
@@ -37,25 +37,6 @@ use std::time::{Duration, Instant};
 use tempfile::TempDir;
 
 const TEST_TEAM: &str = "test-team";
-
-struct LifecycleFlagResetGuard {
-    lifecycle: LifecycleControlSourceAdapter,
-}
-
-impl LifecycleFlagResetGuard {
-    fn install(lifecycle: LifecycleControlSourceAdapter) -> Self {
-        lifecycle.set_terminate_for_test(false);
-        lifecycle.set_reload_for_test(false);
-        Self { lifecycle }
-    }
-}
-
-impl Drop for LifecycleFlagResetGuard {
-    fn drop(&mut self) {
-        self.lifecycle.set_terminate_for_test(false);
-        self.lifecycle.set_reload_for_test(false);
-    }
-}
 
 struct StaleRecoverySignalGuard;
 
@@ -113,23 +94,6 @@ impl RequestDispatcher for DoctorOnlyDispatcher {
     }
 }
 
-#[cfg(unix)]
-fn connect_daemon_local_ipc_until_ready(endpoint_path: &std::path::Path) -> LocalSocketStream {
-    let deadline = Instant::now() + Duration::from_secs(3);
-    loop {
-        match LocalSocketStream::connect(
-            atm_core::protocol::daemon_local_ipc_name_from_path(endpoint_path).expect("ipc name"),
-        ) {
-            Ok(stream) => return stream,
-            Err(error) if Instant::now() < deadline => {
-                let _ = error;
-                std::thread::yield_now();
-            }
-            Err(error) => panic!("connect daemon local ipc: {error}"),
-        }
-    }
-}
-
 #[test]
 fn daemon_shutdown_signals_for_test_are_isolated() {
     let first = LifecycleControlSourceAdapter::new_for_test();
@@ -153,6 +117,30 @@ fn daemon_shutdown_signal_install_reuses_shared_flags() {
 
     assert!(second.reload_requested_for_test());
     assert!(first.terminate_requested());
+}
+
+#[test]
+fn daemon_exit_code_mapping_matches_supervisor_contract() {
+    assert_eq!(
+        daemon_exit_code_for_error(&AtmError::daemon_lifecycle_wedge("wedged")),
+        DaemonExitCode::LifecycleWedge
+    );
+    assert_eq!(
+        daemon_exit_code_for_error(&AtmError::daemon_launch_gate_rejected("launch gate")),
+        DaemonExitCode::DoNotRestart
+    );
+    assert_eq!(
+        daemon_exit_code_for_error(&AtmError::daemon_unavailable("listener died")),
+        DaemonExitCode::TransportFatal
+    );
+    assert_eq!(
+        daemon_exit_code_for_error(&AtmError::config("bad config")),
+        DaemonExitCode::DoNotRestart
+    );
+    assert_eq!(
+        daemon_exit_code_for_error(&AtmError::validation("buggy invariant")),
+        DaemonExitCode::InternalBug
+    );
 }
 
 #[test]
